Add AppFramework/Db base classes
This will allow apps to use Entities that do not have an id attribute.
They can have other unique columns forcing an id is not ideal in all
cases.

For this two base classes are introduced and the current Entity and
QBMapper are actually based on the base classes.

The Entity is really simple and just adds the id property.
The BaseMapper is a bit less clean. As we can't know the where clause
for insert/delete or update parts.
